# config/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Documentação API PetMap",
    "description": "Esta documentação serve como guia para desenvolvedores entenderem como utilizar e integrar a API PetMap de forma eficiente, detalhando endpoints, parâmetros e métodos de autenticação.",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Insira o token no formato: Bearer {token}"
    }
  },
  "paths": {
...
    "/api/notification/create": {
      "post": {
        "tags": [
          "Commentary"
        ],
        "summary": "Criar nova notificação",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/notification/mark-as-read/{notificationId}": {
      "put": {
        "tags": [
          "Commentary"
        ],
        "summary": "Marcar notificação como lida",
        "description": "",
        "parameters": [
          {
            "name": "notificationId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/notification/mark-all-as-read": {
      "put": {
        "tags": [
          "Commentary"
        ],
        "summary": "Marcar todas as notificações do usuário como lidas",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/notification/unread-count": {
      "get": {
        "tags": [
          "Commentary"
        ],
        "summary": "Contar notificações não lidas",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/notification/list": {
      "get": {
        "tags": [
          "Commentary"
        ],
        "summary": "Listar notificações do usuário autenticado",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}